Rename index.ts to transaction.ts
